Disallows string arg for gatway in SlackMessage (#29)
* Disallows string arg for gatway in SlackMessage

This is the final piece of the puzzle. Now that the tests are all
updated to inject the SlackGateway object I can force clients to inject
a SlackGateway object instead of having the SlackMessage class know how
to instantiate that object on its own.

That will allow us to pursue a design that further decouples things.